[TASK] singled out view, added view interface
